+++ src/frontend/apps/impress/src/features/service-worker/service-worker.ts
@@ -185,7 +185,7 @@ registerRoute(
 );
 
 /**
- * Cache stategy static files images (images / svg)
+ * Cache strategy static files images (images / svg)
  */
 registerRoute(
   ({ request }) => request.destination === 'image',
@@ -201,7 +201,7 @@ registerRoute(
 );
 
 /**
- * Cache stategy static files fonts
+ * Cache strategy static files fonts
  */
 googleFontsCache();
 registerRoute(
@@ -218,7 +218,7 @@ registerRoute(
 );
 
 /**
- * Cache stategy static files (css, js, workers)
+ * Cache strategy static files (css, js, workers)
  */
 registerRoute(
   ({ request }) =>